231115 1425 dongjin add today Commit

+++ src/main/java/com/server/pitch/cv/domain/Language.java
@@ -17,6 +17,7 @@ public class Language {
 
     private String exam_type;
     private String language_name;
-    private double language_score;
+    private int language_score;
     private String language_level;
+
 }
